refactor(web): only refresh storage when loading ProposalPage
